move class into it's own file
