+++ HomesEngland.Gateway.Test/AssetGatewayTests.cs
@@ -1,12 +1,8 @@
 ﻿using System;
 using System.Collections;
-using System.Collections.Generic;
 using System.Collections.Immutable;
-using System.Linq;
-using System.Threading;
 using System.Threading.Tasks;
 using System.Transactions;
-using FluentAssertions;
 using HomesEngland.Domain;
 using HomesEngland.Gateway.Migrations;
 using HomesEngland.Gateway.Sql;
@@ -60,55 +56,6 @@ namespace HomesEngland.Gateway.Test
                 var readAsset = await _classUnderTest.ReadAsync(createdAsset.Id).ConfigureAwait(false);
                 //assert
                 readAsset.AssetIsEqual(createdAsset.Id, entity);
-                trans.Dispose();
-            }
-        }
-    }
-
-    [TestFixture]
-    public class BulkAssetCreatorTests
-    {
-        private readonly IGateway<IAsset, int> _gateway;
-        private readonly IBulkAssetCreator _classUnderTest;
-
-        public BulkAssetCreatorTests()
-        {
-            var databaseUrl = Environment.GetEnvironmentVariable("DATABASE_URL");
-            var assetGateway = new EFAssetGateway(databaseUrl);
-            _classUnderTest = assetGateway;
-            _gateway = assetGateway;
-
-            var assetRegisterContext = new AssetRegisterContext(databaseUrl);
-            assetRegisterContext.Database.Migrate();
-        }
-
-        [TestCase(1)]
-        [TestCase(2)]
-        [TestCase(3)]
-        public async Task GivenAnAssetHasBeenCreated_WhenTheAssetIsReadFromTheGateway_ThenItIsTheSame(int count)
-        {
-            //arrange 
-            using (var trans = new TransactionScope(TransactionScopeAsyncFlowOption.Enabled))
-            {
-                
-                IList<IAsset> assets = new List<IAsset>();
-                for (int i = 0; i < count; i++)
-                {
-                    var entity = TestData.Domain.GenerateAsset();
-                    assets.Add(entity);
-                }
-                
-                var createdAssets = await _classUnderTest.BulkCreateAsync(assets, CancellationToken.None).ConfigureAwait(false);
-                //act
-                for (int i = 0; i < count; i++)
-                {
-                    var createdAsset = createdAssets.ElementAtOrDefault(i);
-                    var readAsset = await _gateway.ReadAsync(createdAsset.Id).ConfigureAwait(false);
-                    //assert
-                    createdAsset.Id.Should().NotBe(0);
-                    readAsset.AssetIsEqual(createdAsset.Id, readAsset);
-                }
-                
                 trans.Dispose();
             }
         }
